requirements-prod.txt
# 生产环境完整依赖 (Zeabur 部署使用)
Flask==2.3.3
openpyxl==3.1.2
pytz==2023.3
schedule==1.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask-Session==0.5.0
